% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gads_get_org_units.R
\name{gas_get_org_units}
\alias{gas_get_org_units}
\title{Retrieves a list of all organizational units for an account.}
\usage{
gas_get_org_units(customer_id = "my_customer")
}
\arguments{
\item{customer_id}{The unique ID for the customer's Google Workspace account. As an account administrator, you can also use the my_customer alias to represent your account's customerId. The customerId is also returned as part of the Users resource.}
}
\value{
tibble
}
\description{
Retrieves a list of all organizational units for an account.
}
\examples{
\dontrun{
org_units <- gas_get_org_units()
}
}
