% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gas_get_user_drive_usage.R
\name{gas_get_user_drive_usage}
\alias{gas_get_user_drive_usage}
\title{Retrieves a report which is a collection of properties and statistics for a set of users with the account.}
\usage{
gas_get_user_drive_usage(emails, date = Sys.Date() - 2)
}
\arguments{
\item{emails}{A vector of user emails or "all" to retrieve data for all users.}

\item{date}{The date for which the data should be retrieved (default: two days ago).}
}
\value{
A list containing user drive usage data.
}
\description{
Retrieves a report which is a collection of properties and statistics for a set of users with the account.
}
\details{
For more information, see the \href{https://developers.google.com/admin-sdk/reports/v1/guides/manage-usage-users}{User Usage Report guide}.
}
\examples{
\dontrun{
gas_get_user_drive_usage("all", date = Sys.Date()-2)
}
}
