% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gas_get_user_licenses.R
\name{gas_get_user_licenses}
\alias{gas_get_user_licenses}
\title{List all users assigned licenses for a specific product SKU.}
\usage{
gas_get_user_licenses(product_id = "Google-Apps", customerId = "C00le6y6i")
}
\arguments{
\item{product_id}{A product's unique identifier. For more information about products in this version of the API, see \href{https://developers.google.com/admin-sdk/licensing/v1/how-tos/products}{Products and SKUs}.}

\item{customerId}{The customer's unique ID as defined in the Admin console, such as C00000000. If the customer is suspended, the server returns an error.}
}
\value{
list
}
\description{
List all users assigned licenses for a specific product SKU.
}
\examples{
products <- c(
'Google-Apps',
'101031',
'101037',
'101038',
'Google-Vault',
'101001',
'101005',
'101033',
'101034',
'101047'
)

license <- map_dfr(.x = products, gas_get_user_licenses)
}
