% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gas_get_user_data.R
\name{gas_get_users_data}
\alias{gas_get_users_data}
\title{Retrieves a paginated list of either deleted users or all users in a domain.}
\usage{
gas_get_users_data()
}
\value{
tibble
}
\description{
Retrieves a paginated list of either deleted users or all users in a domain.
}
\examples{
\dontrun{
users_data <- gas_get_users_data()
}
}
