% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gas_auth.R
\name{gas_has_token}
\alias{gas_has_token}
\title{Is there a token on hand?}
\usage{
gas_has_token()
}
\value{
Logical.
}
\description{
Reports whether rgoogleads has stored a token, ready for use in downstream
requests.
}
\seealso{
Other low-level API functions: 
\code{\link{gas_token}()}
}
\concept{low-level API functions}
