% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gas_auth.R
\name{gas_token}
\alias{gas_token}
\title{Produce configured token}
\usage{
gas_token()
}
\value{
A \code{request} object (an S3 class provided by \link[httr:httr-package]{httr}).
}
\description{
For internal use or for those programming around the Google Admin SDK.
Returns a token pre-processed with \code{\link[httr:config]{httr::config()}}. Most users
do not need to handle tokens "by hand" or, even if they need some
control, \code{\link[=gas_auth]{gas_auth()}} is what they need. If there is no current
token, \code{\link[=gas_auth]{gas_auth()}} is called to either load from cache or
initiate OAuth2.0 flow.
If auth has been deactivated via \code{\link[=gas_deauth]{gas_deauth()}}, \code{gas_token()}
returns \code{NULL}.
}
\seealso{
Other low-level API functions: 
\code{\link{gas_has_token}()}
}
\concept{low-level API functions}
