% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gas_get_user_drive_usage.R
\name{get_drive_usage_by_user}
\alias{get_drive_usage_by_user}
\title{Helper for \code{gas_get_user_drive_usage()}}
\usage{
get_drive_usage_by_user(user_email, date = Sys.Date() - 2)
}
\arguments{
\item{user_email}{User primary email or 'all'}

\item{date}{Report date}
}
\value{
tibble
}
\description{
Helper for \code{gas_get_user_drive_usage()}
}
