% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rgoogleadminsdk-package.R
\docType{package}
\name{rgoogleadminsdk-package}
\alias{rgoogleadminsdk}
\alias{rgoogleadminsdk-package}
\title{rgoogleadminsdk: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/selesnow/rgoogleadminsdk}
  \item Report bugs at \url{https://github.com/selesnow/rgoogleadminsdk/issues}
}

}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com}

}
\keyword{internal}
